docs/requirements.txt
sphinx>=7.0
furo
sphinx-autodoc-typehints
nbsphinx
sphinx-click